examples/livekit/pyproject.toml
[project]
name = "voiceobs-livekit-example"
version = "0.1.0"
description = "Example: LiveKit Voice Agent with voiceobs instrumentation"
requires-python = ">=3.9"
dependencies = [
    "voiceobs",
    "livekit-agents[silero,turn-detector]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-deepgram>=1.0.0",
    "livekit-plugins-openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = []

[tool.uv.sources]
voiceobs = { path = "../../" }
